Consolidating YT Json Data into YTDLData.cs.
