+++ Jellyfin.Plugin.YoutubeMetadata/Providers/LocalMetadata/YoutubeLocalEpisodeProvider.cs
@@ -12,7 +12,7 @@ namespace Jellyfin.Plugin.YoutubeMetadata.Providers
 
         public override string Name => Constants.PluginName;
 
-        internal override MetadataResult<Episode> GetMetadataImpl(Utils.YTDLMovieJson jsonObj)
+        internal override MetadataResult<Episode> GetMetadataImpl(Utils.YTDLData jsonObj)
         {
             return Utils.YTDLJsonToEpisode(jsonObj);
         }
